+++ ttoklip-notification/src/main/java/com/notification/config/KafkaPropertiesConfig.java
@@ -16,12 +16,6 @@ public class KafkaPropertiesConfig {
     @Value("${spring.kafka.consumer.group-id}")
     private String groupId;
 
-    @Value("${spring.kafka.consumer.key-deserializer}")
-    private String keyDeserializer;
-
-    @Value("${spring.kafka.consumer.value-deserializer}")
-    private String valueDeserializer;
-
     @Value("${spring.kafka.consumer.enable-auto-commit}")
     private boolean enableAutoCommit;
