+++ ttoklip-notification/src/main/java/com/notification/consumer/NotificationMessage.java
@@ -1,4 +1,4 @@
-package com.notification.config;
+package com.notification.consumer;
 
 import com.common.NotiCategory;
